projects/etims_backup/requirements.txt
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
openpyxl>=3.1.2
pyinstaller>=6.5.0
